perf: widen the header indices to usize
Review asked for this twice and was right twice. The narrow fields kept
Header at 40 bytes rather than 48, and paid for it with three checked
usize to u32 conversions on the spawn path.

Measured, since the 40 bytes was asserted rather than checked: the wider
header is faster. Spawn and await goes 27.20ns to 26.05ns, detached
26.34 to 25.32, and a 512 byte capture 31.92 to 29.67, with
non-overlapping intervals. Across executors it is the same or better at
every count, 26.2 against 27.9 at one and 54.4 against 56.3 at
sixty-four, so there is no cache penalty at scale either.

The conversions cost more than the eight bytes saved, and they could
panic. Both are gone.

+++ glommio/src/executor/multitask.rs
@@ -152,7 +152,7 @@ impl LocalExecutor {
     /// Spawns a thread-local future onto this executor.
     fn spawn<T>(
         &self,
-        executor_id: u32,
+        executor_id: usize,
         tq: Rc<RefCell<TaskQueue>>,
         future: impl Future<Output = T>,
     ) -> (Runnable, JoinHandle<T>) {
@@ -189,7 +189,7 @@ impl LocalExecutor {
 
     pub(crate) fn spawn_and_run<T>(
         &self,
-        executor_id: u32,
+        executor_id: usize,
         tq: Rc<RefCell<TaskQueue>>,
         future: impl Future<Output = T>,
     ) -> Task<T> {
@@ -200,7 +200,7 @@ impl LocalExecutor {
 
     pub(crate) fn spawn_and_schedule<T>(
         &self,
-        executor_id: u32,
+        executor_id: usize,
         tq: Rc<RefCell<TaskQueue>>,
         future: impl Future<Output = T>,
     ) -> Task<T> {
